plataforma/requirements.txt
Django==4.2
django-cors-middleware==1.5.0
django-dotenv==1.4.2
django-grappelli==3.0.8
psycopg2-binary==2.9.5
djangorestframework==3.15.1
djangorestframework-camel-case==1.4.2
django-model-utils==4.4.0
django-rest-knox
django-import-export==4.0.0